Debug Clear errors and omit leaveTrail function
After clear button was clicked, the mouse would still leave a greyscale trail. Change the conditional for greyscale to 'else if' in the changeColor function so that it gives greyscale only if greyscale is clicked.

Squares was defined within a function without let or const. In order to avoid that define squares and add event listeners to them within the createSquares function. For clear button to reach squares include trailColor change also within createSquares function.