Reorganize script to make it more readable
Add a separate event listener for clear button, because it has a different functionality than the other buttons. Define squares constant within the clear function.

Add the event listener for each sqaure to the div element instead of square class and place it in the for loop within the createSquares function, so just after each square is created.

Restructure changeColor function. Create two new functions paintRainbow and paintGreyScale and call those within the changeColor function.