Restructure askNumberOfSquares function
Instead of while loop to ask for a number between 1 and 100 place it in else if so that when the user clicks on cancel after an input outside the range, s/he can turn back to whe s/he was before hitting the popupButton.

Place a separate removePreviousSquares function and call it within askNumberOfSquares function.